Spec improvements
Add a spec that just instantiates full hosts with the right hiera data
etc. to define a mysql server.

just run "bundle exec rake spec:host" to execute it. Currently requires
Ruby 1.9.3, not Ruby 2.1 or 2.2 (due to the version of puppet we're
using not supporting that).

It will actually compile the full catalogue (which means compiling all
the puppet manifests and related functions etc.) - it just doesn't then
attempt to apply it.

Add new relevant specs to spec/hosts/cloud_sql_spec.rb

If you want to test some different hiera data, define a spec for a host
with a new name and create a yaml file with that same name in
spec/fixtures/hiera/